LayoutTests/http/tests/security/webaudio-render-remote-audio-blocked-no-crossorigin-expected.txt

Ensure that audio is not rendered when tainted by a remote audio resource when CORS is not enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS outputArray is silentArray
PASS successfullyParsed is true

TEST COMPLETE
